Separar el main del index. Issue #2
