+++ src/index.html
@@ -13,7 +13,6 @@
   
 
   <!--Desde aquí empiezan los JS-->
-  
 
   <!-- The core Firebase JS SDK is always required and must be listed first -->
   <script src="https://www.gstatic.com/firebasejs/7.6.2/firebase-app.js"></script>
@@ -36,6 +35,6 @@
     // Initialize Firebase
     firebase.initializeApp(firebaseConfig);
   </script>
-  <script type="module" src="main.js"></script>
+  <script type="module" src="main.js"></script>  
 </body>
 </html>
